pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mask-kernel"
version = "0.1.0"
description = "MASK Kernel - Multi-Agent Skill Kit for building expertise agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "MASK Team" },
]
keywords = ["agent", "llm", "skills", "a2a", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "pyyaml>=6.0",
    # CLI
    "typer>=0.12.0",
    "jinja2>=3.1.0",
    # A2A Protocol (with http-server extras for SSE support)
    "a2a-sdk[http-server]>=0.2.0",
    # Server
    "uvicorn>=0.30.0",
    "starlette>=0.38.0",
    "sse-starlette>=2.0.0",
    # HTTP client
    "httpx>=0.27.0",
]

[project.optional-dependencies]
# Storage backends
redis = ["redis>=5.0"]
postgresql = ["asyncpg>=0.29.0", "sqlalchemy>=2.0"]

# Observability - Langfuse (recommended for multi-tenant production)
observability = [
    "langfuse>=2.0.0",
]

# Observability - Phoenix (simpler setup, good for development)
phoenix = [
    "arize-phoenix>=4.0",
    "openinference-instrumentation-langchain>=0.1.0",
]

# LLM Providers
anthropic = ["langchain-anthropic>=0.2.0"]
openai = ["langchain-openai>=0.2.0"]
google = ["langchain-google-genai>=2.0.0"]

# MCP Integration (uses langchain-mcp-adapters)
mcp = ["langchain-mcp-adapters>=0.1.0"]

# Development
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

# All optional dependencies
all = [
    "mask-kernel[redis,postgresql,observability,anthropic,openai,mcp,dev]",
]

[project.scripts]
mask = "mask.cli.main:app"

[project.urls]
Homepage = "https://github.com/colinlee0924/mask-kernel"
Documentation = "https://github.com/colinlee0924/mask-kernel#readme"
Repository = "https://github.com/colinlee0924/mask-kernel"

[tool.hatch.build.targets.wheel]
packages = ["src/mask"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
